Loaded values from db.

+++ src/main/java/ua/vboden/repositories/WordRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ua.vboden.entities.Word;
 
-public interface WordRepository extends CrudRepository<Word, Long> {
+public interface WordRepository extends CrudRepository<Word, Integer> {
 
 }
